Code complete id from uint to long...position from int to long... one failing test remains

+++ src/HatTrick.MemDb/IMemDbAccessor.cs
@@ -6,7 +6,7 @@ namespace HatTrick.InMemDb
     {
         public bool Exists(Func<T, bool> where);
 
-        public bool Exists(uint id);
+        public bool Exists(long id);
 
         public int Count();
 
@@ -14,19 +14,19 @@ namespace HatTrick.InMemDb
 
         public T Find(Func<T, bool> where);
 
-        public T Find(uint id);
+        public T Find(long id);
 
         public T[] FindAll(Func<T, bool> where);
 
-        public T[] FindAll(params uint[] ids);
+        public T[] FindAll(params long[] ids);
 
         public void Insert(T rec, bool encrypt = false);
 
-        public void Insert(T rec, Action<uint> idCallback, bool encrypt = false);
+        public void Insert(T rec, Action<long> idCallback, bool encrypt = false);
 
         public int Update(Action<T> apply, Func<T, bool> where);
 
-        public bool Update(Action<T> apply, uint id);
+        public bool Update(Action<T> apply, long id);
 
         public int Delete(Func<T, bool> where);
